+++ mayan/apps/events/icons.py
@@ -5,9 +5,8 @@ from mayan.apps.appearance.classes import Icon
 icon_event_types_subscriptions_list = Icon(
     driver_name='fontawesome', symbol='list-ol'
 )
-icon_events_list = Icon(driver_name='fontawesome', symbol='list-ol')
 icon_events_for_object = Icon(driver_name='fontawesome', symbol='list-ol')
-icon_events_user_list = Icon(driver_name='fontawesome', symbol='rss')
+icon_events_list = Icon(driver_name='fontawesome', symbol='list-ol')
 icon_object_event_types_user_subcriptions_list = Icon(
     driver_name='fontawesome', symbol='rss'
 )