+++ mayan/apps/events/links.py
@@ -41,9 +41,9 @@ def get_unread_notification_count(context):
     ).filter(read=False).count()
 
 
-link_events_list = Link(
-    icon_class=icon_events_list, permissions=(permission_events_view,),
-    text=_('Events'), view='events:events_list'
+link_current_user_events = Link(
+    icon_class=icon_events_list, text=_('My events'),
+    view='events:current_user_events'
 )
 link_events_details = Link(
     text=_('Events'), view='events:events_list'
@@ -53,6 +53,10 @@ link_events_for_object = Link(
     kwargs=get_kwargs_factory('resolved_object'),
     permissions=(permission_events_view,), text=_('Events'),
     view='events:events_for_object',
+)
+link_events_list = Link(
+    icon_class=icon_events_list, permissions=(permission_events_view,),
+    text=_('Events'), view='events:events_list'
 )
 link_event_types_subscriptions_list = Link(
     icon_class=icon_event_types_subscriptions_list,
@@ -71,10 +75,6 @@ link_object_event_types_user_subcriptions_list = Link(
     kwargs=get_kwargs_factory('resolved_object'),
     permissions=(permission_events_view,), text=_('Subscriptions'),
     view='events:object_event_types_user_subcriptions_list',
-)
-link_user_events = Link(
-    args='resolved_object.pk', text=_('User events'),
-    view='events:user_events'
 )
 link_user_notifications_list = Link(
     badge_text=get_unread_notification_count,