+++ mayan/apps/events/tests/literals.py
@@ -1,0 +1,6 @@
+from __future__ import unicode_literals
+
+TEST_EVENT_TYPE_NAMESPACE_LABEL = 'test event type namespace label'
+TEST_EVENT_TYPE_NAMESPACE_NAME = 'test_event_type_namespace_name'
+TEST_EVENT_TYPE_LABEL = 'test event type label'
+TEST_EVENT_TYPE_NAME = 'test_event_type_name'