+++ mayan/apps/events/tests/test_views.py
@@ -8,15 +8,18 @@ from ..permissions import permission_events_view
 
 
 class EventsViewTestCase(GenericDocumentViewTestCase):
+    auto_upload_document = False
+
     def setUp(self):
         super(EventsViewTestCase, self).setUp()
+        self.test_object = self.test_document_type
 
-        content_type = ContentType.objects.get_for_model(self.test_document)
+        content_type = ContentType.objects.get_for_model(self.test_object)
 
         self.view_arguments = {
             'app_label': content_type.app_label,
             'model': content_type.model,
-            'object_id': self.document.pk
+            'object_id': self.test_object.pk
         }
 
     def _request_events_for_object_view(self):
@@ -27,19 +30,15 @@ class EventsViewTestCase(GenericDocumentViewTestCase):
     def test_events_for_object_view_no_permission(self):
         response = self._request_events_for_object_view()
         self.assertNotContains(
-            response=response, text=self.test_document.label, status_code=403
-        )
-        self.assertNotContains(
-            response=response, text='otal:', status_code=403
+            response=response, text=self.test_object.label, status_code=404
         )
 
     def test_events_for_object_view_with_permission(self):
-        self.grant_permission(permission=permission_events_view)
+        self.grant_access(
+            obj=self.test_object, permission=permission_events_view
+        )
 
         response = self._request_events_for_object_view()
         self.assertContains(
-            response=response, text=self.test_document.label, status_code=200
-        )
-        self.assertNotContains(
-            response=response, text='otal: 0', status_code=200
+            response=response, text=self.test_object.label, status_code=200
         )